20: File Upload & Download

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require("../utils/file");
+
 const { validationResult } = require("express-validator/check");
 
 const Product = require("../models/product");
@@ -16,9 +18,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      errorMessage: "Attached file is not an image.",
+      validationErrors: [],
+    });
+  }
+
+  const imageUrl = image.path;
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -53,21 +73,6 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      // return res.status(500).render("admin/edit-product", {
-      //   pageTitle: "Add Product",
-      //   path: "/admin/add-product",
-      //   editing: false,
-      //   hasError: true,
-      //   product: {
-      //     title: title,
-      //     price: price,
-      //     description: description,
-      //     imageUrl: imageUrl,
-      //   },
-      //   errorMessage: "Database operation failed, please tried again.",
-      //   validationErrors: [],
-      // });
-      // res.redirect("/500");
       const error = new Error("err");
       error.httpStatusCode = 500;
       return next(error);
@@ -106,7 +111,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const updatedImage = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
 
@@ -120,7 +125,6 @@ exports.postEditProduct = (req, res, next) => {
         title: updatedTitle,
         price: updatedPrice,
         description: updatedDesc,
-        imageUrl: updatedImageUrl,
         _id: prodId,
       },
       errorMessage: errors.array()[0].msg,
@@ -135,7 +139,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product
         .save()
         .then((result) => {
@@ -170,7 +177,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
